+++ src/scraper_for_facebook/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from collections.abc import Iterator
 from datetime import date